KVM: selftests: Convert psci_cpu_on_test away from VCPU_ID
Pass around 'struct kvm_vcpu' objects in psci_cpu_on_test instead of
relying on global VCPU_IDs.  Ideally, the test wouldn't have to manually
create vCPUs and thus care about vCPU IDs, but it's not the end of the
world and avoiding that behavior isn't guaranteed to be a net positive
(an attempt at macro shenanigans did not go very well).

Signed-off-by: Sean Christopherson <seanjc@google.com>