KVM: selftests: Sanity check input to automagic ioctls() at build time
Add a static assert to {kvm,vm,cpu}_ioctl() to verify that the size of
the argument provided matches the expected size of the IOCTL.  Because
ioctl() ultimately takes a "void *", it's all too easy to pass in garbage
and not detect the error until runtime.  E.g. during the recent CPUID
rework, selftests happily compiled when vcpu_set_cpuid() unintentionally
passed the cpuid() function as the parameter to ioctl() (a local "cpuid"
parameter was removed, but its use was not replace with "vcpu->cpuid" as
intended).

Tweak a variety of "not technically bugs" issue that trigger the
assertion.

Signed-off-by: Sean Christopherson <seanjc@google.com>