KVM: SVM: Add host SEV-ES save area structure into VMCB via a union
Incorporate the _host_ SEV-ES save area into the VMCB as a union with the
legacy save area.  The SEV-ES variant used to save/load host state is
larger than the legacy save area, but resides at the same offset.  Prefix
the field with "host" to make it as obvious as possible that the SEV-ES
variant in the VMCB is only ever used for host state.  Guest state for
SEV-ES VMs is stored in a completely separate page (VMSA), albeit with
the same layout as the host state.

Add compile-time asserts to ensure the VMCB layout is correct and doesn't
somehow exceed the architectural size of a VMCB (which is also the number
of bytes KVM allocates for a VMCB).

No functional change intended.

Signed-off-by: Sean Christopherson <seanjc@google.com>